Use DNS over HTTPS in Dialing
